JS-18 Integration Tests fixes

+++ sdk-core/src/main/java/io/nem/sdk/model/transaction/MultisigAccountModificationTransactionFactory.java
@@ -27,8 +27,8 @@ import org.apache.commons.lang3.Validate;
 public class MultisigAccountModificationTransactionFactory extends
     TransactionFactory<MultisigAccountModificationTransaction> {
 
-    private final byte minApprovalDelta;
-    private final byte minRemovalDelta;
+    private final int minApprovalDelta;
+    private final int minRemovalDelta;
     private final List<MultisigCosignatoryModification> modifications;
 
 
@@ -51,7 +51,7 @@ public class MultisigAccountModificationTransactionFactory extends
      *
      * @return byte
      */
-    public byte getMinApprovalDelta() {
+    public int getMinApprovalDelta() {
         return minApprovalDelta;
     }
 
@@ -61,7 +61,7 @@ public class MultisigAccountModificationTransactionFactory extends
      *
      * @return byte
      */
-    public byte getMinRemovalDelta() {
+    public int getMinRemovalDelta() {
         return minRemovalDelta;
     }
